requirements.txt
adafruit-blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-framebuf==1.6.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ssd1305==1.3.20
adafruit-circuitpython-typing==1.11.1
adafruit-platformdetect==3.75.0
adafruit-pureio==1.1.11
av==13.1.0
binho-host-adapter==0.1.6
board==1.0
evdev==1.7.1
loguru==0.7.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
picamera2==0.3.21
pidng==4.0.9
piexif==1.1.3
pillow==11.0.0
pyfingerprint==1.5
pyftdi==0.55.4
pyserial==3.5
python-prctl==1.8.1
pyusb==1.2.1
rclone==0.4.4
rpi-gpio==0.7.1
rpi-kms==0.1a1
rpi-libcamera==0.1a6
simplejpeg==1.7.6
sysv-ipc==1.1.0
tqdm==4.67.0
typing-extensions==4.12.2
v4l2-python3==0.3.5
